+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.0 (unreleased)
 
 - Added support for passing literals to distance functions with Sequelize
+- Added support for `halfvec` and `sparsevec` types to Knex.js
 - Dropped support for Node < 18
 
 ## 0.1.8 (2024-02-11)